Throw up some demo-like code

+++ src/gui/guiScrollBar.h
@@ -94,7 +94,7 @@ private:
 
 	void setPosRaw(const s32 &pos);
 	void updatePos();
-	std::optional<s32> target_pos;
+	std::optional<f32> target_pos;
 	u32 last_time_ms = 0;
 	u32 last_delta_ms = 17; // assume 60 FPS
 	void interpolatePos();
